// Front/src/app/interfaces/request/request-auth-google.ts
export interface RequestAuthGoogle {
  isNewUser?: any;
  email?: any;
  family_name?: any;
  given_name?: any;
  name?: any;
  idFirebase?: any;
  locale?: any;
  image?: any;
  verified_email?: any;
  accessToken?: any;
  idToken?: any;
  providerId?: any;
  SignInMethod?: any;
  phoneNumber?: any;
  creationTime?: any;
  lastSignInTime?: any;
}